common -> custom chains: API adoption, preparations for dedicated test file for custom chain creation
